pentru ca cineva nu a fost atent si a pus aiurea niste jsp-uri si un servlet, le-am mutat la locul lor
