+++ ConDr/src/main/webapp/login.html
@@ -38,7 +38,7 @@
                             <button type="submit" style="width: 100%; font-size:1.1em;" class="btn btn-large btn btn-success btn-lg btn-block" ><b>Login</b></button>
 
                         </form>
-                        <form action="/register.jsp">
+                        <form action="/ConDr/src/main/webapp/register.jsp">
                             <button type="submit" style="width: 100%; font-size:1.1em;" class="btn btn-large btn btn-success btn-lg btn-block" ><b>Register</b></button>
                         </form>
